+++ variables/ImportVariable.php
@@ -54,6 +54,16 @@ class ImportVariable
     }
 
     /**
+     * Get viewable asset sources.
+     *
+     * @return array
+     */
+    public function getAssetSources()
+    {
+        return craft()->assetSources->getViewableSources();
+    }
+
+    /**
      * Show history overview.
      *
      * @return array